# BattleShip.Models/Domain/GameStatus.cs
namespace BattleShip.Models.Domain;

public enum GameStatus
{
    Waiting,
    PlacingFleet,
    PlayerTurn,
    ComputerTurn,
    PlayerWon,
    ComputerWon,
    Player1Turn,
    Player2Turn,
    Player1Won,
    Player2Won
}